+++ milvus/MilvusClient.go
@@ -102,6 +102,8 @@ type ConnectParam struct {
 	IPAddress string
 	// Port Server port
 	Port int64
+	// ClientTag client port
+	ClientTag string
 }
 
 // SegmentStat segment statistics