+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ---
 - \#64 - Function HasPartition is not found in go-sdk
 - \#68 - Add "params" structure to process complex json string
+- \#78 - Add interceptor and client_tag metadata
 
 ### Task
 ---